// lib/api/generated/models/utils-case-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const UtilsCaseStatus = {
    CaseStatusOpen: 'open',
    CaseStatusInProgress: 'in_progress',
    CaseStatusResolved: 'resolved',
    CaseStatusClosed: 'closed'
} as const;

export type UtilsCaseStatus = typeof UtilsCaseStatus[keyof typeof UtilsCaseStatus];